Add duke and exception packages
