+++ src/main/java/duke/exception/DukeException.java
@@ -1,9 +1,11 @@
+package duke.exception;
+
 /**
- * The DukeException class represents a custom exception used in the Duke class.
+ * The duke.exception.DukeException class represents a custom exception used in the duke.Duke class.
  */
 public class DukeException extends Exception {
     /**
-     * Constructor for DukeException class.
+     * Constructor for duke.exception.DukeException class.
      *
      * @param msg The error message of the exception.
      */
