+++ src/main/java/duke/exception/EmptyDateException.java
@@ -1,11 +1,13 @@
+package duke.exception;
+
 /**
- * The EmptyDateException class represents a custom exception used in the Duke class.
- * It extends the DukeException class.
+ * The duke.exception.EmptyDateException class represents a custom exception used in the duke.Duke class.
+ * It extends the duke.exception.DukeException class.
  * It is thrown when the date of a deadline or event task is empty.
  */
 public class EmptyDateException extends DukeException{
     /**
-     * Constructor for EmptyDateException class.
+     * Constructor for duke.exception.EmptyDateException class.
      *
      * @param msg The type of task for which the date is left empty.
      */
