+++ src/main/java/duke/exception/EmptyTaskException.java
@@ -1,11 +1,13 @@
+package duke.exception;
+
 /**
- * The EmptyTaskException class represents a custom exception used in the Duke class.
- * It extends the DukeException class.
+ * The duke.exception.EmptyTaskException class represents a custom exception used in the duke.Duke class.
+ * It extends the duke.exception.DukeException class.
  * It is thrown when the description of a task is empty.
  */
 public class EmptyTaskException extends DukeException {
     /**
-     * Constructor for EmptyTaskException class.
+     * Constructor for duke.exception.EmptyTaskException class.
      *
      * @param msg The type of task for which the description is left empty.
      */
